ABX-4062 add notifications to application

+++ src/components/Delete.tsx
@@ -2,18 +2,28 @@ import * as React from 'react'
 import { AppState, View } from '../app'
 import { getActiveWallet } from '../helpers/wallets'
 import { deleteWallet } from '../services/wallet_persistance';
+import * as swal from 'sweetalert'
 
 export class Delete extends React.Component<{appState: AppState, setWalletList: Function, changeView: Function}, {}> {
   constructor (props) {
     super(props)
   }
 
-  public deleteW(accountId: string) {
-    deleteWallet(accountId)
-      .then((wallets) => {
-        this.props.setWalletList(wallets)
-        this.props.changeView(View.create, {})
-      })
+  public async deleteW(accountId: string) {
+    const willDelete = await swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this wallet.",
+      icon: "warning",
+      dangerMode: true,
+      buttons: true
+    })
+    if (willDelete) {
+      deleteWallet(accountId)
+        .then((wallets) => {
+          this.props.setWalletList(wallets)
+          this.props.changeView(View.create, {})
+        })
+    }
   }
 
   render() {
